refector lifecycle db mapping testing needed
