+++ src/main/java/de/samply/bbmri/negotiator/db/util/DbUtil.java
@@ -31,33 +31,21 @@ import java.nio.charset.StandardCharsets;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.sql.Statement;
-import java.sql.Timestamp;
 import java.util.*;
 
 import de.samply.bbmri.negotiator.ConfigFactory;
-import de.samply.bbmri.negotiator.jooq.tables.pojos.*;
-import de.samply.bbmri.negotiator.jooq.tables.pojos.Collection;
 import de.samply.bbmri.negotiator.jooq.tables.records.*;
-import de.samply.bbmri.negotiator.model.*;
-import de.samply.bbmri.negotiator.rest.dto.*;
-import de.samply.bbmri.negotiator.model.QueryCollection;
-import eu.bbmri.eric.csit.service.negotiator.database.DbUtilCollection;
-import eu.bbmri.eric.csit.service.negotiator.database.DbUtilListOfDirectories;
 import eu.bbmri.eric.csit.service.negotiator.database.DbUtilNetwork;
 import eu.bbmri.eric.csit.service.negotiator.mapping.DatabaseListMapper;
 import eu.bbmri.eric.csit.service.negotiator.sync.directory.dto.DirectoryCollection;
 import eu.bbmri.eric.csit.service.negotiator.sync.directory.dto.DirectoryNetworkLink;
 import org.jooq.*;
 import org.jooq.Record;
-import org.jooq.exception.DataAccessException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import de.samply.bbmri.negotiator.Config;
-import de.samply.bbmri.negotiator.NegotiatorConfig;
 import de.samply.bbmri.negotiator.jooq.Tables;
-import de.samply.bbmri.negotiator.jooq.enums.Flag;
-import de.samply.bbmri.negotiator.rest.Directory;
 
 import static org.jooq.impl.DSL.field;
 
@@ -103,122 +91,6 @@ public class DbUtil {
     }
 
 
-    /**
-     * Returns a list of Biobanker's id's who made the sample offers for a given query.
-     * @param config
-     * @param queryId
-     * @return offerMakers
-     */
-    public static List<Integer> getOfferMakers(Config config, int queryId) {
-        List<Integer> offerMakers = config.dsl()
-                .selectDistinct(Tables.OFFER.BIOBANK_IN_PRIVATE_CHAT)
-                .from(Tables.OFFER)
-                .where(Tables.OFFER.QUERY_ID.eq(queryId))
-                .fetch()
-                .getValues(Tables.OFFER.BIOBANK_IN_PRIVATE_CHAT, Integer.class);
-
-        return offerMakers;
-    }
-
-    /**
-     * Returns a list of OfferPersonDTOs for a specific query.
-     * @param config
-     * @param queryId
-     * @return target
-     */
-    public static List<OfferPersonDTO> getOffers(Config config, int queryId, Integer biobankInPrivateChat, int personId) {
-        Result<Record> offerPersons = config.dsl()
-                .select(getFields(Tables.OFFER, "offer"))
-                .select(getFields(Tables.PERSON, "person"))
-                .select(getFields(Tables.PERSON_OFFER, "personoffer"))
-                .from(Tables.OFFER)
-                .join(Tables.PERSON, JoinType.LEFT_OUTER_JOIN).on(Tables.OFFER.PERSON_ID.eq(Tables.PERSON.ID))
-                .leftOuterJoin(Tables.PERSON_OFFER).on(Tables.PERSON_OFFER.OFFER_ID.eq(Tables.OFFER.ID)).and(Tables.PERSON_OFFER.PERSON_ID.eq(personId))
-                .where(Tables.OFFER.QUERY_ID.eq(queryId))
-                .and(Tables.OFFER.BIOBANK_IN_PRIVATE_CHAT.eq(biobankInPrivateChat))
-                .and(Tables.OFFER.STATUS.eq("published"))
-                .orderBy(Tables.OFFER.COMMENT_TIME.asc()).fetch();
-
-        List<OfferPersonDTO> result = new ArrayList<>();
-        HashMap<Integer, List<Collection>> personCollections = new HashMap<>();
-
-        for(Record record : offerPersons) {
-            OfferPersonDTO offerPersonDTO = new OfferPersonDTO();
-            offerPersonDTO.setOffer(config.map(record, Offer.class));
-            offerPersonDTO.getOffer().setId(Integer.parseInt(record.getValue("offer_id").toString()));
-            offerPersonDTO.setPerson(config.map(record, de.samply.bbmri.negotiator.jooq.tables.pojos.Person.class));
-            offerPersonDTO.getPerson().setId(Integer.parseInt(record.getValue("person_id").toString()));
-            offerPersonDTO.setCommentRead(record.getValue("personoffer_read") == null || (boolean) record.getValue("personoffer_read"));
-            Integer commenterId = offerPersonDTO.getPerson().getId();
-            if(!personCollections.containsKey(commenterId)) {
-                Result<Record> collections = config.dsl()
-                        .select(getFields(Tables.COLLECTION, "collection"))
-                        .from(Tables.PERSON_COLLECTION)
-                        .join(Tables.COLLECTION, JoinType.LEFT_OUTER_JOIN).on(Tables.PERSON_COLLECTION.COLLECTION_ID.eq(Tables.COLLECTION.ID))
-                        .join(Tables.QUERY_COLLECTION)
-                            .on(Tables.QUERY_COLLECTION.COLLECTION_ID.eq(Tables.COLLECTION.ID))
-                                .and(Tables.QUERY_COLLECTION.QUERY_ID.eq(queryId))
-                        .where(Tables.PERSON_COLLECTION.PERSON_ID.eq(commenterId))
-                        .fetch();
-                personCollections.put(commenterId, config.map(collections, Collection.class));
-            }
-            offerPersonDTO.setCollections(personCollections.get(commenterId));
-            result.add(offerPersonDTO);
-        }
-
-        return result;
-    }
-
-    public static Result<Record> getCommentCountAndTime(Config config, Integer queryId){
-
-        Result<Record> result = config.dsl()
-                .select(Tables.COMMENT.COMMENT_TIME.max().as("last_comment_time"))
-                .select(Tables.COMMENT.ID.countDistinct().as("comment_count"))
-                .from(Tables.COMMENT)
-                .where(Tables.COMMENT.QUERY_ID.eq(queryId))
-                .and(Tables.COMMENT.STATUS.eq("published"))
-                .fetch();
-
-        return result;
-    }
-
-    public static Result<Record> getPrivateNegotiationCountAndTimeForResearcher(Config config, Integer queryId, Integer personId){
-        Result<Record> result = config.dsl()
-                .select(Tables.OFFER.COMMENT_TIME.max().as("last_comment_time"))
-                .select(Tables.OFFER.ID.countDistinct().as("private_negotiation_count"))
-                .select(Tables.PERSON_OFFER.READ.count().as("unread_private_negotiation_count"))
-                .from(Tables.OFFER)
-                .leftOuterJoin(Tables.PERSON_OFFER)
-                    .on(Tables.PERSON_OFFER.OFFER_ID.eq(Tables.OFFER.ID)
-                            .and(Tables.PERSON_OFFER.PERSON_ID.eq(personId))
-                            .and(Tables.PERSON_OFFER.READ.eq(false)))
-                .where(Tables.OFFER.QUERY_ID.eq(queryId))
-                .and(Tables.OFFER.STATUS.eq("published"))
-                .fetch();
-        return result;
-    }
-
-    public static Result<Record> getPrivateNegotiationCountAndTimeForBiobanker(Config config, Integer queryId, Integer personId){
-        Result<Record> result = config.dsl()
-                .select(Tables.OFFER.COMMENT_TIME.max().as("last_comment_time"))
-                .select(Tables.OFFER.ID.countDistinct().as("private_negotiation_count"))
-                .select(Tables.PERSON_OFFER.READ.count().as("unread_private_negotiation_count"))
-                .select(Tables.COLLECTION.ID.countDistinct().as("number_of_collections"))
-                .from(Tables.OFFER)
-                .join(Tables.BIOBANK).on(Tables.BIOBANK.ID.eq(Tables.OFFER.BIOBANK_IN_PRIVATE_CHAT))
-                .join(Tables.COLLECTION).on(Tables.BIOBANK.ID.eq(Tables.COLLECTION.BIOBANK_ID))
-                .join(Tables.PERSON_COLLECTION).on(Tables.COLLECTION.ID.eq(Tables.PERSON_COLLECTION.COLLECTION_ID))
-                .leftOuterJoin(Tables.PERSON_OFFER)
-                    .on(Tables.PERSON_OFFER.OFFER_ID.eq(Tables.OFFER.ID)
-                        .and(Tables.PERSON_OFFER.PERSON_ID.eq(personId))
-                        .and(Tables.PERSON_OFFER.READ.eq(false)))
-                .where(Tables.OFFER.QUERY_ID.eq(queryId))
-                .and(Tables.OFFER.STATUS.eq("published"))
-                .and(Tables.PERSON_COLLECTION.PERSON_ID.eq(personId)).fetch();
-        return result;
-    }
-
-
     public static String getFullListForAPI(Config config) {
         ResultQuery<Record> resultQuery = config.dsl().resultQuery("SELECT CAST(array_to_json(array_agg(row_to_json(jsond))) AS varchar) AS directories FROM ( " +
                 "SELECT json_build_object('name', d2.name, 'url', d2.url, 'description', d2.description, 'Biobanks', " +
@@ -233,124 +105,13 @@ public class DbUtil {
                 "FROM public.list_of_directories d LEFT JOIN public.list_of_directories d2 ON d2.name = d.directory_prefix " +
                 ") AS jsond;");
         Result<Record> result = resultQuery.fetch();
-        for(Record record : result) {//class org.postgresql.util.PGobject
-
+        for(Record record : result) {
             //PGobject jsonObject = record.getValue(0);
             return (String)record.getValue(0);
         }
         return "ERROR";
     }
 
-    /**
-     * Check if there are queries expecting results from this collection
-     * @param config    DB access handle
-     * @param collectionId    unique id of collection
-     * @return List<QueryCollection> list of qyery_collection records
-     */
-    public static List<de.samply.bbmri.negotiator.model.QueryCollection> checkExpectedResults(Config config, int
-            collectionId){
-        Result<Record2<Integer, Integer>> result = config.dsl()
-                .select(Tables.QUERY_COLLECTION.QUERY_ID, Tables.QUERY_COLLECTION.COLLECTION_ID)
-                .from(Tables.QUERY_COLLECTION)
-                .where(Tables.QUERY_COLLECTION.EXPECT_CONNECTOR_RESULT.eq(true))
-                .and(Tables.QUERY_COLLECTION.COLLECTION_ID.eq(collectionId)).fetch();
-
-        List<QueryCollection> queryCollectionList = config.map(result, QueryCollection.class);
-        return queryCollectionList;
-    }
-
-    /**
-     * Gets the time when the last connector request was made for the negotiations.
-     *
-     * @param config       JOOQ configuration
-     * @param collectionId collection id of the connector
-     * @return Timestamp of last request
-     */
-    public static Timestamp getLastNewNegotiationTime(Config config, String collectionId) {
-        Record1<Timestamp> result = config.dsl()
-                .select(Tables.CONNECTOR_LOG.LAST_NEGOTIATION_TIME)
-                .from(Tables.CONNECTOR_LOG)
-                .where(Tables.CONNECTOR_LOG.DIRECTORY_COLLECTION_ID.eq(collectionId))
-                .and (Tables.CONNECTOR_LOG.LAST_NEGOTIATION_TIME.isNotNull())
-                .orderBy(Tables.CONNECTOR_LOG.LAST_QUERY_TIME.desc())
-                .fetchAny();
-
-        if (result == null) {
-            return null;
-        }
-
-        Timestamp timestamp = result.value1();
-        return timestamp;
-    }
-
-    /**
-     * Gets all the negotiations after the given timestamp.
-     *
-     * @param config    JOOQ configuration
-     * @param timestamp
-     * @return List<QueryDetail> list of queries
-     */
-    public static List<QueryCollection> getAllNewNegotiations(Config config, Timestamp timestamp, String directoryCollectionId) {
-        Integer collectionId = DbUtilCollection.getCollectionId(config, directoryCollectionId);
-
-        Result<Record> result = config.dsl()
-                .select(Tables.QUERY.ID.as("queryId"))
-                .select(Tables.QUERY_COLLECTION.COLLECTION_ID.as("collectionId"))
-                .from(Tables.QUERY)
-                .join(Tables.QUERY_COLLECTION, JoinType.JOIN).on(Tables.QUERY_COLLECTION.QUERY_ID.eq(Tables.QUERY.ID))
-                .where(Tables.QUERY.VALID_QUERY.eq(true))
-                .and(Tables.QUERY.NEGOTIATION_STARTED_TIME.ge(timestamp))
-                .and(Tables.QUERY_COLLECTION.COLLECTION_ID.eq(collectionId))
-                .fetch();
-
-        List<QueryCollection> queryCollectionList = config.map(result, QueryCollection.class);
-        return queryCollectionList;
-    }
-
-
-    /**
-     * Logs the time when the connector request was made for the negotiations.
-     *
-     * @param config                JOOQ configuration
-     * @param directoryCollectionId The collection directoryID
-     */
-    public static void logGetNegotiationTime(Config config, String directoryCollectionId) {
-        //TODO What if foreign key constraint fails
-        ConnectorLogRecord connectorLogRecord = config.dsl().newRecord(Tables.CONNECTOR_LOG);
-        connectorLogRecord.setDirectoryCollectionId(directoryCollectionId);
-        connectorLogRecord.setLastNegotiationTime(new Timestamp(new Date().getTime()));
-        connectorLogRecord.store();
-    }
-
-
-    /**
-     * Gets the time when first negotiation was started in the negotiator.
-     *
-     * @param config JOOQ configuration
-     * @return Timestamp timestamp of negotiation
-     */
-    public static Timestamp getFirstNegotiationTime(Config config) {
-        Record1<Timestamp> result = config.dsl()
-                .select(Tables.QUERY.NEGOTIATION_STARTED_TIME)
-                .from(Tables.QUERY)
-                .where(Tables.QUERY.VALID_QUERY.eq(true))
-                .and (Tables.QUERY.NEGOTIATION_STARTED_TIME.isNotNull())
-                .orderBy(Tables.QUERY.NEGOTIATION_STARTED_TIME.asc())
-                .fetchAny();
-
-        if (result == null) {
-            return null;
-        }
-
-        Timestamp timestamp = result.value1();
-        return timestamp;
-    }
-
-    /**
-     * Executes the given file name as SQL file. It tries to load the file by using the class loader.
-     * @param filename the file name, e.g. "/sql.upgrades/upgrade1.sql"
-     * @throws SQLException
-     */
     public static void executeSQLFile(Connection connection, ClassLoader classLoader, String filename) throws SQLException, IOException {
         InputStream stream = classLoader.getResourceAsStream(filename);
 
@@ -360,11 +121,6 @@ public class DbUtil {
         executeStream(connection, stream);
     }
 
-    /**
-     * Executes the given string as SQL statement.
-     * @param sql the SQL statement
-     * @throws SQLException
-     */
     public static void executeSQL(Connection connection, String sql) throws SQLException {
         try (Statement st = connection.createStatement()) {
             st.execute(sql);
@@ -372,11 +128,6 @@ public class DbUtil {
         }
     }
 
-    /**
-     * Executes the given input stream as SQL statements.
-     * @param stream the input stream for SQL statements.
-     * @throws SQLException
-     */
     public static void executeStream(Connection connection, InputStream stream) throws SQLException, IOException {
         InputStreamReader reader = new InputStreamReader(stream, StandardCharsets.UTF_8);
 
@@ -395,140 +146,6 @@ public class DbUtil {
         }
         br.close();
         executeSQL(connection, sb.toString());
-    }
-
-    /*
-     * Get request staus for lifecycle
-     */
-    public static List<RequestStatusDTO> getRequestStatus(Config config, Integer requestId) {
-        Result<RequestStatusRecord> fetch = config.dsl().selectFrom(Tables.REQUEST_STATUS)
-                .where(Tables.REQUEST_STATUS.QUERY_ID.eq(requestId))
-                .fetch();
-        List<RequestStatusDTO> returnList = new ArrayList<RequestStatusDTO>();
-        for(RequestStatusRecord requestStatusRecord : fetch) {
-            returnList.add(MappingDbUtil.mapRequestStatusDTO(requestStatusRecord));
-        }
-        return returnList;
-    }
-
-    /*
-     * Save request status for lifecycle
-     */
-    public static RequestStatusDTO saveUpdateRequestStatus(Integer requestStatusId, Integer query_id, String status, String status_type, String status_json, Date status_date, Integer status_user_id) {
-        RequestStatusRecord requestStatus = null;
-        try (Config config = ConfigFactory.get()) {
-            if(requestStatusId == null) {
-                requestStatus = config.dsl().newRecord(Tables.REQUEST_STATUS);
-                requestStatus.setQueryId(query_id);
-                requestStatus.setStatus(status);
-                requestStatus.setStatusType(status_type);
-                requestStatus.setStatusJson(status_json);
-                requestStatus.setStatusDate(new Timestamp(status_date.getTime()));
-                requestStatus.setStatusUserId(status_user_id);
-                requestStatus.store();
-                config.commit();
-            } else {
-                config.dsl().update(Tables.REQUEST_STATUS)
-                        .set(Tables.REQUEST_STATUS.QUERY_ID, query_id)
-                        .set(Tables.REQUEST_STATUS.STATUS, status)
-                        .set(Tables.REQUEST_STATUS.STATUS_TYPE, status_type)
-                        .set(Tables.REQUEST_STATUS.STATUS_JSON, status_json)
-                        .set(Tables.REQUEST_STATUS.STATUS_DATE, new Timestamp(status_date.getTime()))
-                        .set(Tables.REQUEST_STATUS.STATUS_USER_ID, status_user_id).where(Tables.REQUEST_STATUS.ID.eq(requestStatusId)).execute();
-                config.commit();
-                requestStatus = config.dsl().selectFrom(Tables.REQUEST_STATUS)
-                        .where(Tables.REQUEST_STATUS.ID.eq(requestStatusId)).fetchOne();
-            }
-            return MappingDbUtil.mapRequestStatusDTO(requestStatus);
-        } catch (SQLException e) {
-            System.err.println("ERROR saving/updating Request Status.");
-            e.printStackTrace();
-        }
-        return null;
-    }
-
-    public static boolean requestStatusForRequestExists(Integer request_id) {
-        try (Config config = ConfigFactory.get()) {
-            int count = config.dsl().selectCount()
-                    .from(Tables.REQUEST_STATUS)
-                    .where(Tables.REQUEST_STATUS.QUERY_ID.eq(request_id))
-                    .fetchOne(0, int.class);
-            return count != 0;
-        } catch (SQLException e) {
-            System.err.println("ERROR saving/updating Request Status.");
-            e.printStackTrace();
-        }
-        return false;
-    }
-
-    public static CollectionRequestStatusDTO saveUpdateCollectionRequestStatus(Integer collectionRequestStatusId, Integer query_id, Integer collection_id, String status, String status_type, String status_json, Date status_date, Integer status_user_id) {
-        QueryLifecycleCollectionRecord collectionRequestStatus = null;
-        try (Config config = ConfigFactory.get()) {
-            if(collectionRequestStatusId == null) {
-                collectionRequestStatus = config.dsl().newRecord(Tables.QUERY_LIFECYCLE_COLLECTION);
-                collectionRequestStatus.setQueryId(query_id);
-                collectionRequestStatus.setCollectionId(collection_id);
-                collectionRequestStatus.setStatus(status);
-                collectionRequestStatus.setStatusType(status_type);
-                collectionRequestStatus.setStatusJson(status_json);
-                collectionRequestStatus.setStatusDate(new Timestamp(status_date.getTime()));
-                collectionRequestStatus.setStatusUserId(status_user_id);
-                collectionRequestStatus.store();
-                config.commit();
-            } else {
-                config.dsl().update(Tables.QUERY_LIFECYCLE_COLLECTION)
-                        .set(Tables.QUERY_LIFECYCLE_COLLECTION.QUERY_ID, query_id)
-                        .set(Tables.QUERY_LIFECYCLE_COLLECTION.COLLECTION_ID, collection_id)
-                        .set(Tables.QUERY_LIFECYCLE_COLLECTION.STATUS, status)
-                        .set(Tables.QUERY_LIFECYCLE_COLLECTION.STATUS_TYPE, status_type)
-                        .set(Tables.QUERY_LIFECYCLE_COLLECTION.STATUS_JSON, status_json)
-                        .set(Tables.QUERY_LIFECYCLE_COLLECTION.STATUS_DATE, new Timestamp(status_date.getTime()))
-                        .set(Tables.QUERY_LIFECYCLE_COLLECTION.STATUS_USER_ID, status_user_id).where(Tables.QUERY_LIFECYCLE_COLLECTION.ID.eq(collectionRequestStatusId)).execute();
-                config.commit();
-                collectionRequestStatus = config.dsl().selectFrom(Tables.QUERY_LIFECYCLE_COLLECTION)
-                        .where(Tables.QUERY_LIFECYCLE_COLLECTION.ID.eq(collectionRequestStatusId)).fetchOne();
-            }
-            return MappingDbUtil.mapCollectionRequestStatusDTO(collectionRequestStatus);
-        } catch (SQLException e) {
-            System.err.println("ERROR saving/updating Request Status.");
-            e.printStackTrace();
-        }
-        return null;
-    }
-
-    public static HashMap<String, String> getOpenRequests() {
-        HashMap<String, String> returnlist = new HashMap<String, String>();
-        try (Config config = ConfigFactory.get()) {
-            ResultQuery<Record> resultQuery = config.dsl().resultQuery("SELECT CASE WHEN status ILIKE 'rejected' THEN 'rejected'\n" +
-                    "WHEN status ILIKE 'under_review' THEN 'review'\n" +
-                    "ELSE 'approved' END statuscase, COUNT(*)\n" +
-                    "\tFROM public.request_status\n" +
-                    "\tWHERE (query_id, status_date) IN (SELECT query_id, MAX(status_date)\n" +
-                    "\tFROM public.request_status GROUP BY query_id) GROUP BY statuscase;");
-            Result<Record> result = resultQuery.fetch();
-            for(Record record : result) {
-                returnlist.put( record.getValue(0).toString(), record.getValue(1).toString() );
-            }
-        } catch (SQLException e) {
-            System.err.println("ERROR getting open Request Status.");
-            e.printStackTrace();
-        }
-        return returnlist;
-    }
-
-    public static List<RequestStatusDTO> getRequestStatusDTOToReview() {
-        try (Config config = ConfigFactory.get()) {
-            Result<Record> fetch = config.dsl().resultQuery("SELECT * FROM public.request_status WHERE query_id IN \n" +
-                    "(SELECT query_id\n" +
-                    "FROM public.request_status\n" +
-                    "WHERE status ILIKE 'under_review' AND (query_id, status_date) IN (SELECT query_id, MAX(status_date)\n" +
-                    "FROM public.request_status GROUP BY query_id) ORDER BY status_date) ORDER BY query_id, status_date;").fetch();
-            return config.map(fetch, RequestStatusDTO.class);
-        } catch (SQLException e) {
-            System.err.println("ERROR getting open Request Status.");
-            e.printStackTrace();
-        }
-        return null;
     }
 
     public static int getNumberOfInitializedQueries() {
