+++ src/main/java/eu/bbmri/eric/csit/service/negotiator/database/DbUtilComment.java
@@ -6,9 +6,11 @@ import de.samply.bbmri.negotiator.db.util.MappingDbUtil;
 import de.samply.bbmri.negotiator.db.util.MappingListDbUtil;
 import de.samply.bbmri.negotiator.jooq.Tables;
 import de.samply.bbmri.negotiator.jooq.tables.pojos.Collection;
+import de.samply.bbmri.negotiator.jooq.tables.pojos.Offer;
 import de.samply.bbmri.negotiator.jooq.tables.records.*;
 import de.samply.bbmri.negotiator.model.CommentAttachmentDTO;
 import de.samply.bbmri.negotiator.model.CommentPersonDTO;
+import de.samply.bbmri.negotiator.model.OfferPersonDTO;
 import de.samply.bbmri.negotiator.model.PrivateAttachmentDTO;
 import eu.bbmri.eric.csit.service.negotiator.database.utils.FieldHelper;
 import eu.bbmri.eric.csit.service.negotiator.mapping.DatabaseListMapper;
@@ -17,6 +19,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.jooq.JoinType;
 import org.jooq.Record;
+import org.jooq.Record1;
 import org.jooq.Result;
 
 import java.sql.SQLException;
@@ -268,5 +271,131 @@ public class DbUtilComment {
                 .fetchOne();
         record.setStatus("deleted");
         record.update();
+    }
+
+    public static List<Integer> getOfferMakers(Config config, int queryId) {
+        List<Integer> offerMakers = config.dsl()
+                .selectDistinct(Tables.OFFER.BIOBANK_IN_PRIVATE_CHAT)
+                .from(Tables.OFFER)
+                .where(Tables.OFFER.QUERY_ID.eq(queryId))
+                .fetch()
+                .getValues(Tables.OFFER.BIOBANK_IN_PRIVATE_CHAT, Integer.class);
+
+        return offerMakers;
+    }
+
+    public static List<OfferPersonDTO> getOffers(Config config, int queryId, Integer biobankInPrivateChat, int personId) {
+        Result<Record> offerPersons = config.dsl()
+                .select(FieldHelper.getFields(Tables.OFFER, "offer"))
+                .select(FieldHelper.getFields(Tables.PERSON, "person"))
+                .select(FieldHelper.getFields(Tables.PERSON_OFFER, "personoffer"))
+                .from(Tables.OFFER)
+                .join(Tables.PERSON, JoinType.LEFT_OUTER_JOIN).on(Tables.OFFER.PERSON_ID.eq(Tables.PERSON.ID))
+                .leftOuterJoin(Tables.PERSON_OFFER).on(Tables.PERSON_OFFER.OFFER_ID.eq(Tables.OFFER.ID)).and(Tables.PERSON_OFFER.PERSON_ID.eq(personId))
+                .where(Tables.OFFER.QUERY_ID.eq(queryId))
+                .and(Tables.OFFER.BIOBANK_IN_PRIVATE_CHAT.eq(biobankInPrivateChat))
+                .and(Tables.OFFER.STATUS.eq("published"))
+                .orderBy(Tables.OFFER.COMMENT_TIME.asc()).fetch();
+
+        List<OfferPersonDTO> result = new ArrayList<>();
+        HashMap<Integer, List<Collection>> personCollections = new HashMap<>();
+
+        for(Record record : offerPersons) {
+            OfferPersonDTO offerPersonDTO = new OfferPersonDTO();
+            offerPersonDTO.setOffer(config.map(record, Offer.class));
+            offerPersonDTO.getOffer().setId(Integer.parseInt(record.getValue("offer_id").toString()));
+            offerPersonDTO.setPerson(config.map(record, de.samply.bbmri.negotiator.jooq.tables.pojos.Person.class));
+            offerPersonDTO.getPerson().setId(Integer.parseInt(record.getValue("person_id").toString()));
+            offerPersonDTO.setCommentRead(record.getValue("personoffer_read") == null || (boolean) record.getValue("personoffer_read"));
+            Integer commenterId = offerPersonDTO.getPerson().getId();
+            if(!personCollections.containsKey(commenterId)) {
+                Result<Record> collections = config.dsl()
+                        .select(FieldHelper.getFields(Tables.COLLECTION, "collection"))
+                        .from(Tables.PERSON_COLLECTION)
+                        .join(Tables.COLLECTION, JoinType.LEFT_OUTER_JOIN).on(Tables.PERSON_COLLECTION.COLLECTION_ID.eq(Tables.COLLECTION.ID))
+                        .join(Tables.QUERY_COLLECTION)
+                            .on(Tables.QUERY_COLLECTION.COLLECTION_ID.eq(Tables.COLLECTION.ID))
+                                .and(Tables.QUERY_COLLECTION.QUERY_ID.eq(queryId))
+                        .where(Tables.PERSON_COLLECTION.PERSON_ID.eq(commenterId))
+                        .fetch();
+                personCollections.put(commenterId, config.map(collections, Collection.class));
+            }
+            offerPersonDTO.setCollections(personCollections.get(commenterId));
+            result.add(offerPersonDTO);
+        }
+
+        return result;
+    }
+
+    //TODO: Harmonice this to use mapping not the resultset
+    public static Result<Record> getPrivateNegotiationCountAndTimeForResearcher(Config config, Integer queryId, Integer personId){
+        Result<Record> result = config.dsl()
+                .select(Tables.OFFER.COMMENT_TIME.max().as("last_comment_time"))
+                .select(Tables.OFFER.ID.countDistinct().as("private_negotiation_count"))
+                .select(Tables.PERSON_OFFER.READ.count().as("unread_private_negotiation_count"))
+                .from(Tables.OFFER)
+                .leftOuterJoin(Tables.PERSON_OFFER)
+                    .on(Tables.PERSON_OFFER.OFFER_ID.eq(Tables.OFFER.ID)
+                            .and(Tables.PERSON_OFFER.PERSON_ID.eq(personId))
+                            .and(Tables.PERSON_OFFER.READ.eq(false)))
+                .where(Tables.OFFER.QUERY_ID.eq(queryId))
+                .and(Tables.OFFER.STATUS.eq("published"))
+                .fetch();
+        return result;
+    }
+
+    //TODO: Harmonice this to use mapping not the resultset
+    public static Result<Record> getPrivateNegotiationCountAndTimeForBiobanker(Config config, Integer queryId, Integer personId){
+        Result<Record> result = config.dsl()
+                .select(Tables.OFFER.COMMENT_TIME.max().as("last_comment_time"))
+                .select(Tables.OFFER.ID.countDistinct().as("private_negotiation_count"))
+                .select(Tables.PERSON_OFFER.READ.count().as("unread_private_negotiation_count"))
+                .select(Tables.COLLECTION.ID.countDistinct().as("number_of_collections"))
+                .from(Tables.OFFER)
+                .join(Tables.BIOBANK).on(Tables.BIOBANK.ID.eq(Tables.OFFER.BIOBANK_IN_PRIVATE_CHAT))
+                .join(Tables.COLLECTION).on(Tables.BIOBANK.ID.eq(Tables.COLLECTION.BIOBANK_ID))
+                .join(Tables.PERSON_COLLECTION).on(Tables.COLLECTION.ID.eq(Tables.PERSON_COLLECTION.COLLECTION_ID))
+                .leftOuterJoin(Tables.PERSON_OFFER)
+                    .on(Tables.PERSON_OFFER.OFFER_ID.eq(Tables.OFFER.ID)
+                        .and(Tables.PERSON_OFFER.PERSON_ID.eq(personId))
+                        .and(Tables.PERSON_OFFER.READ.eq(false)))
+                .where(Tables.OFFER.QUERY_ID.eq(queryId))
+                .and(Tables.OFFER.STATUS.eq("published"))
+                .and(Tables.PERSON_COLLECTION.PERSON_ID.eq(personId)).fetch();
+        return result;
+    }
+
+    public static Timestamp getLastNewNegotiationTime(Config config, String collectionId) {
+        Record1<Timestamp> result = config.dsl()
+                .select(Tables.CONNECTOR_LOG.LAST_NEGOTIATION_TIME)
+                .from(Tables.CONNECTOR_LOG)
+                .where(Tables.CONNECTOR_LOG.DIRECTORY_COLLECTION_ID.eq(collectionId))
+                .and (Tables.CONNECTOR_LOG.LAST_NEGOTIATION_TIME.isNotNull())
+                .orderBy(Tables.CONNECTOR_LOG.LAST_QUERY_TIME.desc())
+                .fetchAny();
+
+        if (result == null) {
+            return null;
+        }
+
+        Timestamp timestamp = result.value1();
+        return timestamp;
+    }
+
+    public static Timestamp getFirstNegotiationTime(Config config) {
+        Record1<Timestamp> result = config.dsl()
+                .select(Tables.QUERY.NEGOTIATION_STARTED_TIME)
+                .from(Tables.QUERY)
+                .where(Tables.QUERY.VALID_QUERY.eq(true))
+                .and (Tables.QUERY.NEGOTIATION_STARTED_TIME.isNotNull())
+                .orderBy(Tables.QUERY.NEGOTIATION_STARTED_TIME.asc())
+                .fetchAny();
+
+        if (result == null) {
+            return null;
+        }
+
+        Timestamp timestamp = result.value1();
+        return timestamp;
     }
 }
