+++ src/main/java/eu/bbmri/eric/csit/service/negotiator/lifecycle/CollectionLifeCycleStatus.java
@@ -2,12 +2,12 @@ package eu.bbmri.eric.csit.service.negotiator.lifecycle;
 
 import de.samply.bbmri.negotiator.Config;
 import de.samply.bbmri.negotiator.ConfigFactory;
-import de.samply.bbmri.negotiator.db.util.DbUtil;
 import de.samply.bbmri.negotiator.jooq.tables.pojos.Person;
 import de.samply.bbmri.negotiator.model.CollectionBiobankDTO;
 import de.samply.bbmri.negotiator.model.CollectionRequestStatusDTO;
 import de.samply.bbmri.negotiator.util.DataCache;
 import eu.bbmri.eric.csit.service.negotiator.database.DbUtilCollection;
+import eu.bbmri.eric.csit.service.negotiator.database.DbUtilLifecycle;
 import eu.bbmri.eric.csit.service.negotiator.lifecycle.requeststatus.*;
 import eu.bbmri.eric.csit.service.negotiator.lifecycle.util.LifeCycleRequestStatusStatus;
 import eu.bbmri.eric.csit.service.negotiator.lifecycle.util.LifeCycleRequestStatusType;
@@ -150,7 +150,7 @@ public class CollectionLifeCycleStatus {
     }
 
     private CollectionRequestStatusDTO createCollectionRequestStatusInDB(String status, String statusType, String status_json, Integer status_user_id) {
-        CollectionRequestStatusDTO collectionRequestStatusDTO = DbUtil.saveUpdateCollectionRequestStatus(null, query_id, collection_id, status, statusType, status_json, new Date(), status_user_id);
+        CollectionRequestStatusDTO collectionRequestStatusDTO = DbUtilLifecycle.saveUpdateCollectionRequestStatus(null, query_id, collection_id, status, statusType, status_json, new Date(), status_user_id);
         return collectionRequestStatusDTO;
     }
